- milestone 2 - Fix webpack problems - Upgrade all packages to latest - Better webpack configurations
