+++ package.json
@@ -7,8 +7,8 @@
     "node": ">=8.9.4 <9.0"
   },
   "scripts": {
-    "start": "cross-env NODE_PATH=. node server.js",
-    "dev": "NODE_ENV=development cross-env NODE_PATH=. node-dev server.js",
+    "start": "webpack --config ./webpack/webpack.prod.js && NODE_PATH=. cross-env node server.js",
+    "dev": "NODE_ENV=development NODE_PATH=. cross-env node server.js",
     "debug": "NODE_ENV=development cross-env NODE_PATH=. node --inspect server.js",
     "setup": "exit 0",
     "mock": "yarn mock-uploader & yarn debug",
@@ -40,23 +40,17 @@
     "@hmcts/uk-bank-holidays": "^1.0.1",
     "accessible-autocomplete": "^1.6.1",
     "applicationinsights": "^1.4.0",
-    "babel-core": "^6.26.3",
-    "babel-loader": "^7.1.5",
-    "babel-preset-env": "^1.7.0",
     "body-parser": "^1.19.0",
     "chalk": "^2.4.2",
     "cookies": "^0.7.3",
-    "copy-webpack-plugin": "^4.0.1",
     "cross-env": "^5.1.3",
-    "css-loader": "^3.0.0",
     "csurf": "^1.10.0",
     "express": "^4.17.1",
     "express-nunjucks": "^2.2.3",
     "extends-classes": "^1.0.5",
-    "file-loader": "^4.0.0",
     "form-data": "^2.3.2",
     "formidable": "^1.2.1",
-    "govuk-frontend": "^2.6.0",
+    "govuk-frontend": "^2.13.0",
     "graceful-fs": "^4.1.15",
     "helmet": "^3.18.0",
     "html-entities": "^1.2.1",
@@ -64,51 +58,63 @@
     "jquery": "^3.4.1",
     "jquery-modal": "^0.9.1",
     "js-base64": "^2.5.1",
-    "libphonenumber-js": "^1.7.18",
+    "libphonenumber-js": "^1.7.19",
     "lodash": "^4.17.4",
-    "mini-css-extract-plugin": "^0.7.0",
-    "mocha-webpack": "^1.1.0",
     "moment": "^2.18.1",
-    "node-sass": "^4.12.0",
-    "nodeman": "^1.1.2",
-    "nunjucks-loader": "^3.0.0",
-    "postcss-loader": "^3.0.0",
     "request": "^2.87.0",
     "request-promise": "^4.2.4",
-    "sass-loader": "^7.1.0",
     "serve-static": "^1.14.1",
-    "superagent": "^5.0.6",
-    "webpack": "^4.3.3",
-    "webpack-dev-middleware": "^3.7.0",
-    "webpack-node-externals": "^1.7.2"
+    "superagent": "^5.1.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.4.5",
+    "@babel/preset-env": "^7.4.5",
     "@hmcts/eslint-config": "^1.4.0",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^8.0.6",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "chai-jq": "^0.0.9",
+    "chokidar": "^3.0.1",
     "codeceptjs": "^2.1.4",
+    "copy-webpack-plugin": "^5.0.3",
+    "css-loader": "^3.0.0",
     "eslint": "^5.16.0",
     "eslint-plugin-mocha": "^5.0.0",
+    "file-loader": "^4.0.0",
     "istanbul": "^1.1.0-alpha.1",
     "jsdom": "^15.1.1",
+    "mini-css-extract-plugin": "^0.7.0",
     "mocha": "^6.1.4",
     "mocha-multi": "^1.1.0",
-    "mochawesome": "^4.0.0",
+    "mocha-webpack": "^1.1.0",
+    "mochawesome": "^4.0.1",
     "nightmare": "^3.0.2",
     "nightmare-upload": "^0.1.1",
     "nock": "^10.0.6",
     "node-dev": "^4.0.0",
+    "node-sass": "^4.12.0",
+    "nodemon": "^1.19.1",
+    "nunjucks-loader": "^3.0.0",
     "nyc": "^14.1.1",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
     "pa11y": "^5.1.0",
     "portscanner": "^2.1.1",
+    "postcss-loader": "^3.0.0",
     "pre-commit": "^1.2.2",
     "proxyquire": "^2.0.1",
     "puppeteer": "^1.17.0",
+    "sass-loader": "^7.1.0",
     "sinon": "^7.3.2",
     "sinon-chai": "^3.0.0",
     "sonar-scanner": "^3.0.3",
     "supertest": "^4.0.2",
-    "webdriverio": "^5.10.0"
+    "terser-webpack-plugin": "^1.3.0",
+    "webdriverio": "^5.10.4",
+    "webpack": "^4.34.0",
+    "webpack-cli": "^3.3.4",
+    "webpack-dev-middleware": "^3.7.0",
+    "webpack-merge": "^4.2.1",
+    "webpack-node-externals": "^1.7.2"
   }
 }
