+++ webpack.prod.js
@@ -1,0 +1,11 @@
+const merge = require('webpack-merge');
+const common = require('./webpack.common.js');
+const TerserJSPlugin = require('terser-webpack-plugin');
+const OptimizeCSSAssetsPlugin = require('optimize-css-assets-webpack-plugin');
+
+module.exports = merge(common, {
+  mode: 'production',
+  optimization: {
+    minimizer: [new TerserJSPlugin({}), new OptimizeCSSAssetsPlugin({})]
+  }
+});
